+++ hyunjina/.env
@@ -1,0 +1,1 @@
+VITE_SERVICE_KEY = 1adb8ca5-cc8b-4687-a614-177f3a3b914b
