+++ hyunjina/week1-2/src/App.tsx
@@ -1,35 +1,69 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
+import { useEffect, useState } from "react";
+import axios from "axios";
+import { parseStringPromise } from "xml2js";
 
-function App() {
-  const [count, setCount] = useState(0)
-
-  return (
-    <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
-  )
+interface ShowItem {
+  prfnm: string;
+  fcltynm: string;
+  prfpdfrom: string;
+  prfpdto: string;
+  poster: string;
 }
 
-export default App
+type RawItem = {
+  prfnm: string[];
+  fcltynm: string[];
+  prfpdfrom: string[];
+  prfpdto: string[];
+  poster: string[];
+};
+
+const SERVICE_KEY = import.meta.env.VITE_SERVICE_KEY;
+
+function App() {
+  const [shows, setShows] = useState<ShowItem[]>([]);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const API_URL = `http://www.kopis.or.kr/openApi/restful/pblprfr?service=${SERVICE_KEY}&stdate=20240101&eddate=20241231&cpage=1&rows=5`;
+
+      try {
+        const response = await axios.get(API_URL, { responseType: "text" });
+        const result = await parseStringPromise(response.data);
+        const items = result.dbs.db;
+
+        const parsed: ShowItem[] = (items as RawItem[]).map((item) => ({
+          prfnm: item.prfnm[0],
+          fcltynm: item.fcltynm[0],
+          prfpdfrom: item.prfpdfrom[0],
+          prfpdto: item.prfpdto[0],
+          poster: item.poster[0],
+        }));
+
+        setShows(parsed);
+      } catch (error) {
+        console.error("API 호출 실패:", error);
+      }
+    };
+
+    fetchData();
+  }, []);
+
+  return (
+    <div style={{ padding: "20px" }}>
+      <h1>🎭 공연 목록</h1>
+      {shows.map((show, idx) => (
+        <div key={idx} style={{ marginBottom: "20px" }}>
+          <img src={show.poster} alt={show.prfnm} style={{ width: "150px" }} />
+          <h2>{show.prfnm}</h2>
+          <p>📍 장소: {show.fcltynm}</p>
+          <p>
+            🗓️ 기간: {show.prfpdfrom} ~ {show.prfpdto}
+          </p>
+        </div>
+      ))}
+    </div>
+  );
+}
+
+export default App;
